+++ CHANGELOG.md
@@ -2,6 +2,7 @@ In Development
 --------------
 - Added a page listing the most depended-on projects
 - Added a description for the `flake8_import_order.styles` entry point group
+- Lower the logging level of most messages from `purge-old-versions`
 
 v2020.3.18
 ----------